Switch to a simpler storage model.
Going back to the previous storage model of where we only store the messages.
We reference records by appending the offset of the record to the end of the
uri.

When a tag is discovered we start the service to save it. A PendingIntent
is created that will resolve to the currently running activity.

Change-Id: I32d1dd70960fdf67129d1707399d667de030ebe5