Hack in tag writing support for non-user builds.
The UX is horrible, but it does let you do
basic tag writing. To use it go to My Tag
create a tag you want to write the select
"Write to next tag scanned" from the context
menu of the tag you want to write.

The next time any tag is scanned it will try
to write that message and will show a toast
with the result of the write attempt.

Regardless of success the write request will
be erased at the scan.

Change-Id: Ib4eb4762979bf178bfdb588b7d8a6788b02631d2