Force tags to be in portrait.
This fixes a bug where rotations would drop the record.
Unfortunately, while fixing this, I discovered that the UI design just
doesn't work in landscape. There's a list halfway at the bottom of the
UI, and you can't put a ListView in a ScrollView, so the user has no way
of scrolling down to see the list contents.

Change-Id: Ic282c066789831ff02725895112e884ea7864c07